requirements.txt
langgraph>=0.0.15
pydantic>=2.0.3
pandas>=2.0.0
python-dotenv>=1.0.0
PyYAML>=6.0
requests>=2.31.0
tqdm>=4.64.1
openai>=1.3.0
prefect>=2.0.0